CoreCracker_v1/requirements.txt
annotated-types    0.7.0
bcrypt             4.3.0
certifi            2025.1.31
cffi               1.17.1
charset-normalizer 3.4.1
cryptography       44.0.2
et_xmlfile         2.0.0
funkify            0.4.5
future             1.0.0
greenlet           3.1.1
idna               3.10
netmiko            4.3.0
ntc_templates      7.7.0
openpyxl           3.0.10
paramiko           3.5.1
pip                25.0.1
psycopg            3.2.1
psycopg-binary     3.2.1
pycparser          2.22
pydantic           2.7.4
pydantic_core      2.18.4
pydantic-settings  2.3.4
PyNaCl             1.5.0
pyserial           3.5
PySocks            1.7.1
python-dotenv      1.1.0
PyYAML             6.0.2
requests           2.32.0
scp                0.15.0
six                1.17.0
SQLAlchemy         2.0.34
sshtunnel          0.1.5
textfsm            1.1.3
typing_extensions  4.13.1
tzdata             2025.2
urllib3            2.3.0
xtyping            0.8.2